requirements.txt
fastapi
uvicorn
requests
pandas
numpy
scikit-learn
pytz
python-dotenv
